+++ app_core/app/database/migrations/2015_02_05_143026_create_detalle_material_colocados_table.php
@@ -20,7 +20,7 @@ class CreateDetalleMaterialColocadosTable extends Migration {
 		Schema::create('detalle_material_colocado', function(Blueprint $table)
 		{
 			$table->increments('id');
-			$table->integer('cantidad');
+			$table->decimal('cantidad', 10, 2);
 			$table->boolean('reempleo');
 			
 			$table->integer('hoja_diaria_id')->unsigned();
